+++ .storybook/stories/Tag.stories.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import '/src/styles/idsk3_theme.css';
 import { Tag } from '../../src/components/Atoms';
@@ -12,20 +12,37 @@ export default {
 const Template: ComponentStory<typeof Tag> = (args) => <Tag {...args} />;
 export const Static = Template.bind({});
 Static.args = {
-  label: 'Static tag'
+  label: 'Static tag',
+  type: 'static'
 };
 
-export const Intractable = Template.bind({});
-Intractable.args = {
-  label: 'You can hover this',
-  interaction: true
+export const Select = Template.bind({});
+Select.args = {
+  label: 'Selectable tag',
+  type: 'select'
 };
 
-export const Selected = Template.bind({});
-Selected.args = {
+export const Action = Template.bind({});
+Action.args = {
+  label: 'Action tag',
+  type: 'action'
+};
+
+export const Filter = Template.bind({});
+Filter.args = {
   label: 'Selected tag',
-  interaction: true,
-  selected: true
+  type: 'filter'
+};
+
+export const Custom = Template.bind({});
+Custom.args = {
+  label: 'Custom colors',
+  type: 'static',
+  colors: {
+    color: '#fc0384',
+    background: '#b8a3c9',
+    border: '#5328ed'
+  }
 };
 
 export const Disabled = Template.bind({});