+++ src/components/Atoms/DropDown/DropDown.tsx
@@ -1,15 +1,10 @@
-import React, {
-  useState,
-  useRef,
-  Children,
-  useEffect,
-  ReactElement,
-  SVGProps,
-  ReactNode
-} from 'react';
+import React, { useState, useRef, Children, ReactElement, SVGProps, ReactNode } from 'react';
 import classNames from 'classnames';
 
 import { ArrowDropDownIcon } from '../../../svgIcons/Navigation';
+import { useClickOutside } from '../../../utils';
+import BaseButton from '../Button/BaseButton';
+import useDropdownDirection from '../../../utils/useDropdownDirection';
 
 export interface DropDownProps extends React.AllHTMLAttributes<HTMLDivElement> {
   id?: string;
@@ -20,6 +15,7 @@ export interface DropDownProps extends React.AllHTMLAttributes<HTMLDivElement> {
   buttonAriaLabel?: string;
   optionsSide?: 'left' | 'right';
   closeOnOptionClick?: boolean;
+  withoutPseudoElement?: boolean;
 }
 
 const DropDown = ({
@@ -33,33 +29,27 @@ const DropDown = ({
   buttonAriaLabel,
   optionsSide = 'right',
   closeOnOptionClick = false,
+  withoutPseudoElement = false,
   ...props
 }: DropDownProps) => {
   const [opened, setOpened] = useState<boolean>(false);
 
   const containerRef = useRef<HTMLDivElement>(null);
+  const optionsRef = useRef<HTMLUListElement>(null);
 
-  useEffect(() => {
-    function handleClickOutside(event: MouseEvent | null) {
-      if (
-        event?.target instanceof Node &&
-        containerRef.current &&
-        !containerRef.current.contains(event.target)
-      ) {
-        setOpened(false);
-      }
-    }
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [containerRef]);
+  const direction = useDropdownDirection(opened, containerRef, optionsRef);
+
+  useClickOutside(() => {
+    setOpened(false);
+  }, containerRef);
 
   const optionClasses = classNames(
     'idsk-dropdown__options',
     {
       hidden: !opened,
-      'idsk-dropdown__options--left': optionsSide === 'left'
+      'idsk-dropdown__options--left': optionsSide === 'left',
+      'idsk-dropdown__options--up': direction === 'up',
+      'idsk-dropdown__options--down': direction === 'down'
     },
     optionClassName
   );
@@ -68,13 +58,20 @@ const DropDown = ({
 
   const renderedChildren = Children.map(children, (child) => {
     if (React.isValidElement(child)) {
+      const label = child.props.label || child.props.children;
+      const data =
+        !withoutPseudoElement && child.type !== 'hr' && !!label
+          ? { 'data-pseudolabel': label }
+          : {};
+      console.log(child.type);
       return (
-        <li className="idsk-dropdown__option">
+        <li className="idsk-dropdown__option idsk-pseudolabel__wrapper" {...data}>
           {React.cloneElement(child, {
             onClick: (e: React.MouseEvent) => {
               if (child?.props?.onClick) child.props.onClick(e);
               if (closeOnOptionClick) setOpened(false);
-            }
+            },
+            className: classNames(child.props.className, 'absolute')
           } as any)}
         </li>
       );
@@ -87,7 +84,7 @@ const DropDown = ({
 
   return (
     <div ref={containerRef} {...props} className={wrapperClasses}>
-      <button
+      <BaseButton
         aria-label={buttonAriaLabel}
         className={buttonClasses}
         onClick={() => setOpened((p) => !p)}
@@ -98,8 +95,9 @@ const DropDown = ({
       >
         <span>{dropDownTitle}</span>
         {renderedIcon}
-      </button>
+      </BaseButton>
       <ul
+        ref={optionsRef}
         id={id + '-dropdown'}
         className={optionClasses}
         data-testid="dropdown-options"