+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as formatBytes } from './formatBytes';
 export { default as useForwardRef } from './useForwardRef';
+export { default as useClickOutside } from './useClickOutside';